Moved DataAccess from database to use_case package as that is where it belongs in CA. Added a comment in Javadoc to HumanPlayerFactory and removed todo comment.
