+++ src/main/java/use_cases/DataAccess.java
@@ -1,4 +1,4 @@
-package database;
+package use_cases;
 
 import entities.HumanPlayer;
 import usecases.PlayerInformation;
